fix: change ms from uint32 to uint64

+++ platform/os/freertos/HAL_Timer_freertos.c
@@ -43,7 +43,7 @@ extern "C" {
 #include "stm32l4xx_hal.h"
 #endif
 
-uint32_t HAL_GetTimeMs(void)
+uint64_t HAL_GetTimeMs(void)
 {
 #if defined PLATFORM_HAS_TIME_FUNCS
     struct timeval tv;
@@ -83,7 +83,7 @@ char *HAL_Timer_current(char *time_str)
 
 bool HAL_Timer_expired(Timer *timer)
 {
-    uint32_t now_ts;
+    uint64_t now_ts;
 
     now_ts = HAL_GetTimeMs();
